requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
replicate==0.32.1
openai==1.58.1
Pillow==11.0.0
huggingface-hub==0.27.0
